requirements.txt
Authlib~=0.15.6
apscheduler==3.11
boto3~=1.36.22 
Bcrypt-Flask==1.0.2
bcrypt==4.3.0
click-option-group~=0.5.5
connexion[swagger-ui]~=2.14.2
cryptography>=44.0.1
dnspython==2.7.0
dramatiq[redis,watch]==1.18.0
email-validator~=2.2.0
eventlet==0.40.3
flask-cors==6.0.0
flask-login~=0.6.2
flask-marshmallow~=0.14.0
flask-restful==0.3.10
flask-shell-ipython==0.5.1
flask-wtf~=1.2.0
Flask-APScheduler==1.13.1
Flask-Mail~=0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==2.5.1
Flask~=2.2.5
flask-socketio==5.6.0
gevent-websocket==0.10.1
giturlparse~=0.12.0
gunicorn~=23.0.0
itsdangerous~=2.2.0
jwt==1.3.1
jsonschema==4.23.0
loginpass==0.5
marshmallow-sqlalchemy~=0.28.2
marshmallow==3.26.1
networkx~=3.3
nf-core~=2.5.0
pika~=1.3.2
prometheus-flask-exporter~=0.22
psycopg2-binary~=2.9.6
pygit2~=1.17.0
pyopenssl==25.0.0
pytest~=8.3.5
pytest-mock~=3.14.0
pytest-rerunfailures~=15.1
pytest-xdist~=3.6.1
python-dotenv~=1.0.1
python-jenkins==1.8.3
python-redis-lock~=4.0.0
PyGithub==1.55
PyYAML~=6.0.1
redis~=6.4.0
repo2rocrate~=0.1.2
requests~=2.32.3
rich~=13.9.4
rocrate==0.13.0
snakemake==7.32.4
roc-validator==0.6.2
SQLAlchemy==1.4.41
urllib3==2.6.3
uwsgi==2.0.28
watchdog==6.0.0
wheel~=0.46.2
# Werkzeug~=2.2.x
git+https://github.com/pallets/werkzeug.git@2.2.x#egg=werkzeug
